backend/requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
sqlalchemy
